refactor: reorganise components into 'section' folder
- Moved all components (Hero, Projects, Contact, About, Skills) from 'src/components/*' to 'src/components/section/*'.
- Updated all imports across related test files to reflect the new folder structure.
