+++ src/app/page.tsx
@@ -1,8 +1,8 @@
-import Hero from "@/components/Hero/Hero";
-import About from "@/components/About/About";
-import Projects from "@/components/Projects/Projects";
-import Skills from "@/components/Skills/Skills";
-import Contact from "@/components/Contact/Contact";
+import Hero from "@/components/section/Hero/Hero";
+import About from "@/components/section/About/About";
+import Projects from "@/components/section/Projects/Projects";
+import Skills from "@/components/section/Skills/Skills";
+import Contact from "@/components/section/Contact/Contact";
 
 export default function Home() {
   return (
