+++ src/components/section/About/About.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import About from "@/components/About/About";
+import About from "@/components/section/About/About";
 
 jest.mock('@/data/personalInterests', () => ({
   personalInterests: [
